clean up a lot and modularize more
